+++ apps/web/jest.config.js
@@ -5,9 +5,15 @@ module.exports = {
     '^.+\\.(ts|tsx)$': 'ts-jest',
     '^.+\\.(js|jsx)$': 'babel-jest',
   },
+  globals: {
+    'ts-jest': {
+      tsconfig: { jsx: 'react-jsx' },
+    },
+  },
   moduleNameMapper: {
     '^react$': '<rootDir>/node_modules/react',
     '^react-dom$': '<rootDir>/node_modules/react-dom',
     '^@/(.*)$': '<rootDir>/src/$1',
   },
+  testPathIgnorePatterns: ['/node_modules/', '/e2e/'],
 };