+++ apps/web/src/components/Layout.tsx
@@ -1,11 +1,13 @@
-import React from 'react';
+import React, { ReactNode } from 'react';
 import Sidebar from './Sidebar';
+import NotLoggedInAlert from './NotLoggedInAlert';
 
-const Layout = ({ children }) => {
+const Layout = ({ children }: { children: ReactNode }) => {
   return (
     <div className="flex h-screen bg-blend-darken">
       <Sidebar />
       <div className="flex-grow p-6">{children}</div>
+      <NotLoggedInAlert />
     </div>
   );
 };